+++ testing/test_telemetry.py
@@ -35,4 +35,4 @@ def test_telemetry_signalk():
 def test_telemetry_nmea0183():
     msg = Nmea0183Messages()
     msg.set_reading(test_reading)
-    assert msg.render() == "$IIVWR,206.0,,,N,5.69,M,,K"
+    assert msg.render() == "$IIVWR,154.0,L,,N,5.69,M,,K"