+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 ===========
 - Add software tests for telemetry subsystem
 - NMEA-0183: Fix message structure
+- NMEA-0183: Fix wind direction and add "left/right of bow" indicator
 
 
 2022-07-14 0.1.0